browser-extension: add BulkSavePage route and UI; refactor SavePage; remove close-tabs toggle from Options; background routing and settings enrichment
