+++ apps/browser-extension/src/main.tsx
@@ -6,6 +6,7 @@ import { HashRouter, Route, Routes } from "react-router-dom";
 
 import BookmarkDeletedPage from "./BookmarkDeletedPage.tsx";
 import BookmarkSavedPage from "./BookmarkSavedPage.tsx";
+import BulkSavePage from "./BulkSavePage.tsx";
 import Layout from "./Layout.tsx";
 import NotConfiguredPage from "./NotConfiguredPage.tsx";
 import OptionsPage from "./OptionsPage.tsx";
@@ -21,6 +22,7 @@ function App() {
           <Routes>
             <Route element={<Layout />}>
               <Route path="/" element={<SavePage />} />
+              <Route path="/bulk-save" element={<BulkSavePage />} />
               <Route
                 path="/bookmark/:bookmarkId"
                 element={<BookmarkSavedPage />}
